+++ cmake/nfxGraphicsSources.cmake
@@ -13,6 +13,7 @@ set(NFX_GRAPHICS_SOURCES
     src/gl/pipeline/passes/GeometryPass.cpp
     src/gl/pipeline/passes/GridPass.cpp
     src/gl/pipeline/passes/PointShadowPass.cpp
+    src/gl/pipeline/passes/PostProcessPass.cpp
     src/gl/pipeline/passes/PresentPass.cpp
     src/gl/pipeline/passes/SpotShadowPass.cpp
     src/gl/pipeline/passes/TransparentPass.cpp